+++ crypto/fipsmodule/bcm.c
@@ -130,6 +130,9 @@
 #include "kem/kem.c"
 #include "md4/md4.c"
 #include "md5/md5.c"
+#ifdef ENABLE_DILITHIUM
+#include "ml_dsa/ml_dsa.c"
+#endif
 #include "ml_kem/ml_kem.c"
 #include "modes/cbc.c"
 #include "modes/cfb.c"
